+++ src/main/java/shedar/mods/ic2/nuclearcontrol/gui/GuiRemoteThermo.java
@@ -74,7 +74,7 @@ public class GuiRemoteThermo extends GuiContainer {
             heat += delta;
             if (heat < 0) heat = 0;
             if (heat >= 1000000) heat = 1000000;
-            if (container.remoteThermo.getHeatLevel().intValue() != heat) {
+            if (container.remoteThermo.getHeatLevel() != heat) {
                 ((NetworkManager) IC2.network.get()).initiateClientTileEntityEvent(container.remoteThermo, heat);
             }
